+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - event delegation. You can use it in this way: `layer.on('click', 'Circle', handler);`
+- new `node.findAncestors(selector)` and `node.findAncestor(selector)` functions
 
 ### Changed
 - `moveTo` and some other methods return `this`